+++ src/fides/core/config/logging_settings.py
@@ -7,7 +7,7 @@ from logging import CRITICAL, DEBUG, ERROR, INFO, WARNING, getLevelName
 from pydantic import validator
 
 from .fides_settings import FidesSettings
-from .utils import get_test_mode
+from .utils import get_dev_mode
 
 ENV_PREFIX = "FIDES__LOGGING__"
 
@@ -34,7 +34,7 @@ class LoggingSettings(FidesSettings):
     def validate_log_level(cls, value: str) -> str:
         """Ensure the provided LEVEL is a valid value."""
 
-        if get_test_mode():
+        if get_dev_mode():
             return getLevelName(DEBUG)
 
         valid_values = [